Comments and minifixes

+++ Shared/Interop/TimingPacket.cs
@@ -4,25 +4,33 @@
     {
 
         //properties for easier jsoning
-        public int originCycle { get; set; }
+        public long originCycle { get; set; }
 
         public List<PinInstruction> instructions { get;} = [];
 
-        public TimingPacket(int originCycle) { 
+        public TimingPacket(long originCycle) { 
             this.originCycle = originCycle;
         }
 
-        public TimingPacket(int originCycle, params PinInstruction[] inst) {
+        public TimingPacket(long originCycle, params PinInstruction[] inst) {
             this.originCycle=originCycle;
             instructions.AddRange(inst);
             instructions.Sort();
         }
-
+        /// <summary>
+        /// Adds an instruction to the list and resorts it.
+        /// </summary>
+        /// <param name="inst">The instruction to add.</param>
+        /// <returns>The added instruction (for chaining purposes).</returns>
         public PinInstruction addInstruction(PinInstruction inst) { 
             instructions.Add(inst);
             instructions.Sort();
             return inst;
         }
+        /// <summary>
+        /// Adds a variable number of unchained premade instructions at once. Currently unused.
+        /// </summary>
+        /// <param name="inst">The instructions to add.</param>
         public void addInstructions(params PinInstruction[] inst)
         {
             instructions.AddRange(inst);
@@ -55,10 +63,15 @@
             cumulUsSinceOriginCycle = prev.cumulUsSinceOriginCycle + furtherUsDelay;
         }
 
-
+        /// <summary>
+        /// Defines custom compare so that sorting works correctly in the timing packet.
+        /// Based on the microseconds since originating cycle.
+        /// </summary>
+        /// <param name="other">another instruction to compare against.</param>
+        /// <returns>0 if they execute at the same time, -1 or 1 as necessary to sort lowest to highest.</returns>
         public int CompareTo(PinInstruction? other)
         {
-            //avoid double precision errors
+            //avoid double precision errors on equals
             if ((cumulUsSinceOriginCycle - other?.cumulUsSinceOriginCycle) < .001)
             {
                 return 0;
